used downsample_array in median_filtered_max_projection
this involves altering some of the logic in downsample_array so that
it uses np.ceil() when determining how many bins to divide an array
into
