+++ src/ophys_etl/utils/array_utils.py
@@ -4,6 +4,33 @@ import numpy as np
 from typing import Union
 
 
+def n_frames_from_hz(
+        input_frame_rate: float,
+        downsampled_frame_rate: float) -> int:
+    """
+    Find the number of frames to group together to downsample
+    a video from input_frame_rate to downsampled_frame_rate
+
+    Parameters
+    ----------
+    input_frame_rate: float
+
+    downsampled_frame_rate: float
+
+    Returns
+    -------
+    frames_to_group: int
+
+    Notes
+    -----
+    If input_frame_rate/downsampled_frame_rate < 1, will return 1
+    """
+
+    frames_to_group = np.round(input_frame_rate/downsampled_frame_rate)
+    frames_to_group = frames_to_group.astype(int)
+    return max(1, frames_to_group)
+
+
 def downsample_array(
         array: Union[h5py.Dataset, np.ndarray],
         input_fps: float = 31.0,
@@ -37,8 +64,8 @@ def downsample_array(
         raise ValueError("downsampling with strategy 'maximum' is not defined")
 
     npts_in = array.shape[0]
-    npts_out = int(npts_in * output_fps / input_fps)
-    bin_list = np.array_split(np.arange(npts_in), npts_out)
+    frames_to_group = n_frames_from_hz(input_fps, output_fps)
+    npts_out = max(1, np.ceil(npts_in/frames_to_group).astype(int))
 
     array_out = np.zeros((npts_out, *array.shape[1:]))
 
@@ -54,8 +81,10 @@ def downsample_array(
             }
 
     sampler = sampling_strategies[strategy]
-    for i, bin_indices in enumerate(bin_list):
-        array_out[i] = sampler(array, bin_indices)
+    for i_out, i0 in enumerate(range(0, npts_in, frames_to_group)):
+        i1 = min(npts_in, i0+frames_to_group)
+        array_out[i_out] = sampler(array,
+                                   np.arange(i0, i1, dtype=int))
 
     return array_out
