+++ tests/utils/test_array_utils.py
@@ -4,6 +4,20 @@ from ophys_etl.utils import array_utils as au
 
 
 @pytest.mark.parametrize(
+        "input_frame_rate, downsampled_frame_rate, expected",
+        [(22.0, 50.0, 1),
+         (100.0, 25.0, 4),
+         (100.0, 7.0, 14),
+         (100.0, 8.0, 12),
+         (100.0, 7.9, 13)])
+def test_n_frames_from_hz(
+        input_frame_rate, downsampled_frame_rate, expected):
+    actual = au.n_frames_from_hz(input_frame_rate,
+                                 downsampled_frame_rate)
+    assert actual == expected
+
+
+@pytest.mark.parametrize(
         ("array, input_fps, output_fps, random_seed, strategy, expected"),
         [
             (
@@ -55,6 +69,12 @@ from ophys_etl.utils import array_utils as au
                 7, 2, 0, 'average',
                 np.array([[13/4, 4], [19/3, 22/3]])),
             (
+                # average downsampel ND array with only 1 output frame
+                np.array([[1, 2], [3, 4], [5, 6]]),
+                10, 1, 0, 'average',
+                np.array([[3.0, 4.0]])
+            ),
+            (
                 # maximum downsample 1D array
                 np.array([1, 4, 6, 2, 3, 5, 11]),
                 7, 2, 0, 'maximum',
@@ -97,6 +117,30 @@ def test_downsample_exceptions(array, input_fps, output_fps, random_seed,
                 output_fps=output_fps,
                 strategy=strategy,
                 random_seed=random_seed)
+
+
+@pytest.mark.parametrize('input_fps', [3, 4, 5])
+def test_decimate_video(input_fps):
+    """
+    This is another test of downsample array to make sure that
+    it treats video-like arrays the way our median_filtered_max_projection
+    code expects
+    """
+    rng = np.random.default_rng(62134)
+    video = rng.random((71, 40, 40))
+
+    expected = []
+    for i0 in range(0, 71, input_fps):
+        frame = np.mean(video[i0:i0+input_fps, :, :],
+                        axis=0)
+        expected.append(frame)
+    expected = np.array(expected)
+
+    actual = au.downsample_array(video,
+                                 input_fps=input_fps,
+                                 output_fps=1,
+                                 strategy='average')
+    np.testing.assert_array_equal(expected, actual)
 
 
 @pytest.mark.parametrize(
